deploy/agentcore/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
bedrock-agentcore
fastmcp>=2.0.0
boto3>=1.28.0
rich>=13.0.0
click>=8.0.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
aws-ops-agent
